Cleanup
Former-commit-id: 017edfb316b799ea3c6204c9a67f530de268138b [formerly 017edfb316b799ea3c6204c9a67f530de268138b [formerly 6c35b5c478dcc2f72756dd627dd007abacc5bf4e]]
Former-commit-id: ec6f12324622fe2bc4bbac2cf46899347f2c5108
Former-commit-id: f8e4e09737d43b7a1c064c7d233a21afb095cae1

+++ openquake/baselib/general.py
@@ -28,6 +28,7 @@ import copy
 import math
 import socket
 import random
+import atexit
 import zipfile
 import operator
 import warnings
@@ -343,12 +344,14 @@ def writetmp(content=None, dir=None, prefix="tmp", suffix="tmp"):
     return path
 
 
+@atexit.register
 def removetmp():
     """
     Remove the temporary files created by writetmp
     """
     for path in _tmp_paths:
-        os.remove(path)
+        if os.path.exists(path):  # not removed yet
+            os.remove(path)
 
 
 def git_suffix(fname):
